Updated Dev

+++ app/Http/Controllers/User/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\User;
 
 use App\Http\Controllers\Controller;
+use App\Http\Controllers\Features\ActivityController;
 use App\Models\profile;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -18,7 +19,7 @@ class UserController extends Controller
         if (Auth::check()) {
             $user = Auth::user();
             $profiles = Profile::where('id', '<>', $user->profile->id)->get();
-            return view('dashboard.user.admin.admin_tables.user_table', compact('profiles'));
+            return view('components.tables.user_table', compact('profiles'));
         } else {
             return redirect()->route('user.login');
         }
@@ -55,12 +56,13 @@ class UserController extends Controller
             return redirect()->back()->with('fail', 'Unable to Create User!');
         }
         $user = new User();
-        $user->name = $request->input('name');
+
         $user->email = $request->input('email');
         $user->password = Hash::make('12345678');
         $user->role = $request->input('role');
         $user->save();
         $profile = new Profile();
+        $profile->name = $request->input('name');
         $profile->age = $request->input('age');
         $profile->gender = $request->input('gender');
         $profile->position = $request->input('position');
@@ -72,18 +74,17 @@ class UserController extends Controller
         }
         $profile->user_id = $user->id;
         $user->profile()->save($profile);
-        return redirect()->back()->with('success', 'Form submitted successfully.');
-    }
-    //Get Specific User
-    public function GetUser($id)
-    {
-        if (Auth::check()) {
-            $user = User::with('profile')->find($id);
-            dd($user);
-            return view('dashboard.user.admin.admin_tables.user_table', compact('user'));
-        } else {
-            return redirect()->route('user.login');
+
+        $activityRequest = new Request([
+            'activity' => 'User',
+            'activity_title' => $user->email,
+            'activity_type' => 'Add',
+        ]);
+        $Activity = new ActivityController();
+        if ($Activity->addActivity($activityRequest)) {
+            return redirect()->back()->with('success', 'Form submitted successfully.');
         }
+        return redirect()->back()->with('error', 'Unable to add User');
     }
     //To EDIT User
     public function updateUser(Request $request)
@@ -103,42 +104,53 @@ class UserController extends Controller
         if (!$findUser) {
             return redirect()->route('user.home')->with('fail', 'User not found!');
         }
-        $findUser->user()->update([
-            'name' => $findUser->user->name,
-            'email' => $findUser->user->email,
-            'role' => $validate['role']
+        $activityRequest = new Request([
+            'activity' => 'User',
+            'activity_title' => $findUser->user->email,
+            'activity_type' => 'Update',
         ]);
-        $findUser->update([
-            'age' => $validate['age'],
-            'gender' => $validate['gender'],
-            'position' => $validate['position'],
-            'department' => $validate['department'],
-            'phone_number' => $validate['phone_number'],
-        ]);
-        return redirect()->back()->with('success', 'User updated successfully.');
+        $Activity = new ActivityController();
+        if ($Activity->addActivity($activityRequest)) {
+            $findUser->user()->update([
+                'email' => $findUser->user->email,
+                'role' => $validate['role']
+            ]);
+            $findUser->update([
+                'age' => $validate['age'],
+                'gender' => $validate['gender'],
+                'position' => $validate['position'],
+                'department' => $validate['department'],
+                'phone_number' => $validate['phone_number'],
+            ]);
+            return redirect()->back()->with('success', 'User updated successfully.');
+        }
     }
     //To Delete User
-    public function deleteUser(Request $request, $id)
+    public function deleteUser(Request $request)
     {
         $validate = $request->validate([
             'id' => 'required',
             'imagesURL' => 'required'
         ]);
-        // Get the image filename or identifier you want to delete
         $imageFilename = $validate;
-        // Determine the storage path of the image
         $storagePath = 'public/images/' . $imageFilename;
-
-        // Delete the image file
-        Storage::delete($storagePath);
         $user = User::find($validate['id']);
         $profile = Profile::find($validate['id']);
-        if ($user && $profile) {
-            // Perform the deletion
-            $user->delete();
-            $profile->delete();
-            return redirect()->back()->with('success', 'User deleted successfully!');
+        $activityRequest = new Request([
+            'activity' => 'User',
+            'activity_title' => $user->email,
+            'activity_type' => 'Delete',
+        ]);
+        $Activity = new ActivityController();
+        if ($Activity->addActivity($activityRequest)) {
+            if ($user && $profile) {
+                // Perform the deletion
+                Storage::delete($storagePath);
+                $user->delete();
+                $profile->delete();
+                return redirect()->back()->with('success', 'User deleted successfully!');
+            }
+            return redirect()->back()->with('fail', 'Failed to delete the user.');
         }
-        return redirect()->back()->with('fail', 'Failed to delete the user.');
     }
 }
